[SYCL][Graph] Unify more tests
- Unify more tests
- Remove some unncessary tests
- Replace DEPENDS_ON macro with templated function
