+++ sycl/test-e2e/Graph/Explicit/host_task.cpp
@@ -9,87 +9,6 @@
 // Expected fail as host tasks aren't implemented yet.
 // XFAIL: *
 
-// This test uses a host_task when explicitly adding a command_graph node.
+#define GRAPH_E2E_EXPLICIT
 
-#include "../graph_common.hpp"
-
-int main() {
-  queue Queue;
-
-  using T = int;
-
-  if (!Queue.get_device().has(sycl::aspect::usm_shared_allocations)) {
-    return 0;
-  }
-
-  const T ModValue = T{7};
-  std::vector<T> DataA(Size), DataB(Size), DataC(Size);
-
-  std::iota(DataA.begin(), DataA.end(), 1);
-  std::iota(DataB.begin(), DataB.end(), 10);
-  std::iota(DataC.begin(), DataC.end(), 1000);
-
-  std::vector<T> Reference(DataC);
-  for (unsigned n = 0; n < Iterations; n++) {
-    for (size_t i = 0; i < Size; i++) {
-      Reference[i] += (DataA[i] + DataB[i]) + ModValue + 1;
-    }
-  }
-
-  exp_ext::command_graph Graph{Queue.get_context(), Queue.get_device()};
-
-  T *PtrA = malloc_device<T>(Size, Queue);
-  T *PtrB = malloc_device<T>(Size, Queue);
-  T *PtrC = malloc_shared<T>(Size, Queue);
-
-  Queue.copy(DataA.data(), PtrA, Size);
-  Queue.copy(DataB.data(), PtrB, Size);
-  Queue.copy(DataC.data(), PtrC, Size);
-  Queue.wait_and_throw();
-
-  // Vector add to output
-  auto NodeA = Graph.add([&](handler &CGH) {
-    CGH.parallel_for(range<1>(Size),
-                     [=](item<1> id) { PtrC[id] += PtrA[id] + PtrB[id]; });
-  });
-
-  // Modify the output values in a host_task
-  auto NodeB = Graph.add(
-      [&](handler &CGH) {
-        CGH.host_task([=]() {
-          for (size_t i = 0; i < Size; i++) {
-            PtrC[i] += ModValue;
-          }
-        });
-      },
-      {exp_ext::property::node::depends_on(NodeA)});
-
-  // Modify temp buffer and write to output buffer
-  Graph.add(
-      [&](handler &CGH) {
-        CGH.parallel_for(range<1>(Size), [=](item<1> id) { PtrC[id] += 1; });
-      },
-      {exp_ext::property::node::depends_on(NodeB)});
-
-  auto GraphExec = Graph.finalize();
-
-  event Event;
-  for (unsigned n = 0; n < Iterations; n++) {
-    Event = Queue.submit([&](handler &CGH) {
-      CGH.depends_on(Event);
-      CGH.ext_oneapi_graph(GraphExec);
-    });
-  }
-  Queue.wait_and_throw();
-
-  Queue.copy(PtrC, DataC.data(), Size);
-  Queue.wait_and_throw();
-
-  free(PtrA, Queue);
-  free(PtrB, Queue);
-  free(PtrC, Queue);
-
-  assert(Reference == DataC);
-
-  return 0;
-}
+#include "../Inputs/host_task.cpp"
