+++ sycl/test-e2e/Graph/Explicit/repeated_exec.cpp
@@ -6,52 +6,6 @@
 //
 // CHECK-NOT: LEAK
 
-// Test executing a graph created with the explicit API multiple times.
+#define GRAPH_E2E_EXPLICIT
 
-#include "../graph_common.hpp"
-
-int main() {
-
-  queue Queue{gpu_selector_v};
-
-  exp_ext::command_graph Graph{Queue.get_context(), Queue.get_device()};
-
-  const size_t N = 10;
-  float *Arr = malloc_device<float>(N, Queue);
-  float ZeroPattern = 0.0f;
-  Queue.fill(Arr, ZeroPattern, N).wait();
-
-  Graph.add([&](handler &CGH) {
-    CGH.parallel_for(range<1>{N}, [=](id<1> idx) {
-      size_t i = idx;
-      Arr[i] += 1;
-    });
-  });
-
-  std::vector<float> Output(N);
-  Queue.memcpy(Output.data(), Arr, N * sizeof(float)).wait();
-  for (int i = 0; i < N; i++)
-    assert(Output[i] == 0);
-
-  auto ExecGraph = Graph.finalize();
-
-  Queue.memcpy(Output.data(), Arr, N * sizeof(float)).wait();
-  for (int i = 0; i < N; i++)
-    assert(Output[i] == 0);
-
-  Queue.submit([&](handler &CGH) { CGH.ext_oneapi_graph(ExecGraph); }).wait();
-
-  Queue.memcpy(Output.data(), Arr, N * sizeof(float)).wait();
-  for (int i = 0; i < N; i++)
-    assert(Output[i] == 1);
-
-  Queue.submit([&](handler &CGH) { CGH.ext_oneapi_graph(ExecGraph); }).wait();
-
-  Queue.memcpy(Output.data(), Arr, N * sizeof(float)).wait();
-  for (int i = 0; i < N; i++)
-    assert(Output[i] == 2);
-
-  sycl::free(Arr, Queue);
-
-  return 0;
-}
+#include "../Inputs/repeated_exec.cpp"
